+++ api/src/main/java/com/grash/repository/PreventiveMaintenanceRepository.java
@@ -23,5 +23,5 @@ public interface PreventiveMaintenanceRepository extends JpaRepository<Preventiv
 
     @Query("SELECT CASE WHEN COUNT(p) > :threshold THEN true ELSE false END " +
             "FROM PreventiveMaintenance p WHERE p.company.id = :companyId")
-    boolean hasMoreThan(@Param("companyId") Long companyId, @Param("threshold") int threshold);
+    boolean hasMoreThan(@Param("companyId") Long companyId, @Param("threshold") Long threshold);
 }
